Add actors in series details
