+++ Models/Movies.cs
@@ -20,6 +20,5 @@ namespace MvcMovie.Models{
         public IdentityUser User { get; set; }
         public string UserId { get; set; }
         public List<Watchlist> Watchlist { get; set; }
-        public List<MvcActor.Models.Actor> Actor {get; set;}
     }
 }
