+++ Views/Series/Details.cshtml
@@ -1,6 +1,5 @@
 @model dynamic
 @using Microsoft.AspNetCore.Identity
-
 @inject SignInManager<IdentityUser> SignInManager
 @inject UserManager<IdentityUser> UserManager
 
@@ -50,35 +49,44 @@
         <dd class="col-sm-10">
             @Model.Series.Overview
         </dd>
+        <dt>
+            @foreach (var actor in Model.Actor)
+                {
+                    @if (actor != null)
+                    {
+                        @actor.FirstName @actor.LastName
+                    }
+                }
+        </dt>
         @foreach (var comment in Model.SComments)
         {
-            <dt class="col-sm-2">
+            <dd class="col-sm-2">
             Comment creator
-            </dt>
-            <dd class="col-sm-10">
+            </dd>
+            <dt class="col-sm-10">
                 @comment.Creator
-            </dd>
-            <dt class="col-sm-2">
+            </dt>
+            <dd class="col-sm-2">
                 Comment Date
-            </dt>
-            <dd class="col-sm-10">
+            </dd>
+            <dt class="col-sm-10">
                 @comment.PubDate
-            </dd>
-            <dt class="col-sm-2">
-                Comment Text
             </dt>
-            <dd class="col-sm-10">
-                @comment.Text
+            <dd class="col-sm-2">
+                Comment Text
             </dd>
+            <dt class="col-sm-10">
+                @comment.Text
+            </dt>
         }
-        <dt class="col-sm-2">
+        <dd class="col-sm-2">
             @if(SignInManager.IsSignedIn(User)){
                 <form asp-action="NewComment" asp-controller="Series" asp-route-id="@Model.Series.Id">
                 <label for="text">Your comment: </label>
                 <input type="text" name="text">
                 </form>
             }
-        </dt>
+        </dd>
     </dl>
     <div>
     <a asp-action="Index">Back to List</a>
